requirements.txt
apns-client==0.2.1
cffi==1.7.0
cryptography==1.4
falcon==1.0.0
gunicorn==19.6.0
idna==2.1
paho-mqtt==1.2
pkg-resources==0.0.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
python-mimeparse==1.5.2
requests==2.11.1
six==1.10.0
